Add missed refactoring changes
