+++ src/main/java/frc/robot/subsystems/shooter/ShooterSubsystem.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.shooter;
 
 import java.lang.Math;
